feat: overhaul media parsing, dispatch to Anki; marked for MD applied
Added setting to toggle between base64 transmission of media files to
Anki or use the absolute path for it.
Save the media files to adhere to the `addNote`-like json schema which
enables the media file to be sent when the note is created of its fields
are updated. For this to work I had to patch AnkiConnect - not sure if
this actually is accepted by upstream now.

Also always dispatch media files to Anki since currently there's no way
to determine changes in local media files from the ones in Anki...

Accidentally amended this giant commit with an even more giant one:

- Restructured `main.ts` and settings
- Created real explorative test files for the vault with a appropriate
  structure
- Made note models actually updatable - which can be triggered manually
- Extracted CSS & scripts to file & folders that are automatically
- Restructure the architecture from Spring Boot-y => oroper OOP with
  invariants for the classes
- Use a new setting value to track the version of the note models on
  which host at which plugin version
- Introduce shiki for hljs-less source-code parsing
  - Might be the reason for the exploded loading-time & bundle size >.<
- Overhauled MathJax parsing with placeholder substitution to fix
  previous bugs like * => <it> etc.
- Many things more I forgot about (?)